Removed default Websocket server implementations, Kestrel are now the default
